Container component created and renaming on all components files as 'index.js'
