+++ portfolio/src/components/main/Container/index.js
@@ -1,0 +1,11 @@
+function Container({ styles, children }) {
+	return (
+		<>
+			<div className="container" style={styles}>
+				{children}
+			</div>
+		</>
+	);
+}
+
+export default Container;
